Fix #20: conserje multi-tenant via /[orgSlug]/conserje (resuelve org desde slug)
